Example 1 has a primative amount of merged parameter stuff in, enough for it to expose bugs
If Parameters are just another Tick -> Number transformer
We need to deliver all ticks and get all the numbers and context mutations in one hop. We used zip for this
However, if one stream calls onComplete, the whole thing should be onCompleted. There is some weirdness about the order of delivery. The example 1 prints enough debug to understand, although determining the exact problem is difficult